requirements.txt
django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
gunicorn==20.0.4
django-cors-headers==3.7.0